app/requirements.txt
Flask
Redis
python-dotenv
pandas
matplotlib

